+++ CHANGELOG.md
@@ -26,6 +26,9 @@ The format is based on [Keep a Changelog].
 - Minimum linting for some mistakes that reduces the avatar performance `#1549`
   - AAO now performs basic linting to identify common mistakes that can negatively impact avatar performance.
   - Currently, multi-pass rendering with exactly the same material is detected since it's likely a mistake that clicks '+' button on the inspector by mistake.
+- Basic mesh support for remove mesh components `#1530`
+  - You now can remove some portion of basic meshes with Remove Mesh components!
+  - This does not includes remove mesh by blendshape because basic mesh does not support blendshape.
 
 ### Changed
 - Avatar Optimizer will run as late as possible in NDMF Pipeline by default `#1493`